refactor: folder structure, simplifying code, animation
Refactored code
Fix folder structure
Add loader animation
