+++ src/components/RecommendationCard/index.scss
@@ -1,0 +1,65 @@
+.cardWrapper {
+  border-radius: 0.25rem;
+  box-shadow: 0 1px 3px 0 rgba(101, 119, 134, 0.7);
+  margin-top: 1rem;
+  padding: 1.75rem 3rem 4.125rem;
+    
+  h2{
+    font-size: 1.25rem;
+    font-weight: 500;
+    margin-bottom: 2.5rem;
+  }
+
+  .center-container {
+    width: 100%;
+    
+    .loader {
+      margin-top: 0;
+    }
+  }
+}
+.recommenderWrapper {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
+
+.mentorWrapper {
+  flex: 1;
+
+  .mentorContent {
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+  }
+
+  img {
+    border-radius: 4px;
+    max-width: 90px;
+    max-height: 90px;
+  }
+
+  .mentorText {
+    margin-left: 15px;
+    position: relative;
+    bottom: 5px;
+  }
+    
+  h2 {
+    font-size: 1.25rem;
+    margin: 0;
+    font-weight: 500;
+  }
+
+  h3 {
+    font-size: 1rem;
+    font-weight: normal;
+    margin: 0;
+  }  
+}
+
+@media (max-width: 800px)   {
+  .cardWrapper {
+    display: none;
+  }
+}
