--- src/screens/recommendation/index.module.scss
@@ -1,60 +1,0 @@
-$h2_fontSize:1.25rem;
-
-.cardWrapper{
-    border-radius: 4px;
-    box-shadow: 0 1px 3px 0 rgba(101, 119, 134, 0.7);
-    margin-top: 1rem;
-    padding: 30px 48px 66px;
-    
-    h2{
-        font-size: $h2_fontSize;
-        font-weight: 500;
-        margin-bottom:40px ;
-    }
-}
-.recommenderWrapper{
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-
-.mentorWrapper{
-    flex: 1;
-    .mentorContent{
-        display: flex;
-        flex-direction: row;
-        width: 100%;
-    }
-
-    img{
-        border-radius: 4px;
-        max-width: 90px;
-        max-height: 90px;
-
-    } 
-
-    .mentorText{
-        margin-left: 15px;
-        position: relative;
-        bottom: 5px;
-    }
-    
-    h2{
-        font-size: $h2_fontSize;
-        margin: 0;
-        font-weight: 500;
-    }
-
-    h3{
-        font-size: 1rem;
-        font-weight: normal;
-        margin: 0;
-    }
-    
-}
-
-@media (max-width: 800px){
-    .cardWrapper{
-        display: none;
-    }
-}
